+++ R/discover.r
@@ -44,6 +44,55 @@
 #'   true, and x = NA being false for any non-NA value of x.
 #' }
 #'
+#' ## Floating-point variables
+#'
+#' Numerical/complex values, i.e. floating-point values, represent difficulties
+#' for stating functional dependencies. A fundamental condition for stating
+#' functional dependencies is that we can compare two values for the same
+#' variable, and they are equivalent or not equivalent.
+#'
+#' Usually, this is done by checking they're equal -- this is the approach used
+#' in \code{discover} -- but we can use any comparison that is an equivalence
+#' relation.
+#'
+#' However, checking floating-point values for equality is not simple. \code{==}
+#' is not appropriate, even when comparing non-calculated values we've read from
+#' a file, because how a given number is converted into a float can vary by
+#' computer architecture, meaning that two values can be considered equal on one
+#' computer, and not equal on another. This can happen even if they're both
+#' using 64-bit R, and even though all R platforms work with values conforming
+#' to the same standard (see \code{\link{double}}). For example,
+#' \eqn{8.54917750000000076227} and \eqn{8.54917749999999898591} are converted into
+#' different floating-point representations on x86, but the same representation
+#' on ARM, resulting in inequality and equality respectively.
+#'
+#' For this and other reasons, checking numerical/complex values for
+#' (near-)equality in R is usually done with \code{\link{all.equal}}. This
+#' determines values \eqn{x} and \eqn{y} to be equal if their absolute/relative
+#' absolute difference is within some tolerance value. However, we can not use
+#' this. Equivalence relations must be transitive: if we have values \eqn{x},
+#' \eqn{y}, and \eqn{z}, and \eqn{x} is equivalent to both \eqn{y} and \eqn{z},
+#' then \eqn{y} and \eqn{z} must also be equivalent. This tolerance-based
+#' equivalence is not transitive: it is reasonably straightforward to set up
+#' three values so that the outer values are far enough apart to be considered
+#' non-equivalent, but the middle value is close enough to be considered
+#' equivalent to both of them. Using this to determine functional dependencies,
+#' therefore, could easily result in a large number of inconsistencies.
+#'
+#' This means we have no good option for comparing numerical/complex values
+#' as-is for equivalence, with consistent results across different machines, so
+#' we must treat them differently. We have three options:
+#'
+#' - Round/truncate the values, before comparison, to some low degree of precision;
+#' - Coerce the values to another class before passing them into \code{discover};
+#' - Read values as characters if reading data from a file.
+#'
+#' \code{discover} takes the first option, with a default number of significant
+#' digits low enough to ensure consistency across different machines. However,
+#' the user can also use any of these options when processing the data before
+#' passing it to \code{discover}. The third option, in particular, is
+#' recommended if reading data from a file.
+#'
 #' ## Skipping bijections
 #'
 #' Skipping bijections allows skipping redundant searches. For example, if the
@@ -101,6 +150,13 @@
 #' @param df a data.frame, the relation to evaluate.
 #' @param accuracy a numeric in (0, 1]: the accuracy threshold required in order
 #'   to conclude a dependency.
+#' @param digits a positive integer, indicating how many significant digits are
+#'   to be used for numeric and complex variables. A value of \code{NA} results
+#'   in no rounding. By default, this uses \code{getOption("digits")}, similarly
+#'   to \code{\link{format}}. See the "Floating-point variables" section below
+#'   for why this rounding is necessary for consistent results across different
+#'   machines. See the note in \code{\link{print.default}} about \code{digits >=
+#'   16}.
 #' @param full_cache a logical, indicating whether to store information about
 #'   how sets of attributes group the relation records (stripped partitions).
 #'   Otherwise, only the number of groups is stored. Storing the stripped
@@ -160,6 +216,7 @@
 discover <- function(
   df,
   accuracy,
+  digits = getOption("digits"),
   full_cache = TRUE,
   store_cache = TRUE,
   skip_bijections = FALSE,
@@ -209,8 +266,23 @@ discover <- function(
 
   # convert all columns to integers, since they're checked for duplicates more
   # quickly when calculating partitions
+  # we must round floating-point/complex columns, since they're otherwise
+  # infeasible:
+  # - all.equal, i.e. equality by tolerance, isn't transient, so isn't an
+  #   equivalence relation; we need such a relation for consistent partitioning
+  # - ==, identical, etc., i.e. equality by bit comparison, results in different
+  #   results on different machines, e.g. x86 and ARM(?) 64-bit both represent
+  #   floats in 64-bit, but x86 represents in 80 bits first and then rounds,
+  #   so non-representable numbers get approximated differently, resulting in
+  #   different partition results
+  if (!is.na(digits))
+    report$exp(
+      df[] <- lapply(df, format_if_float, digits = digits),
+      paste("formatting numerical/complex variables with", digits, "significant digits")
+    )
   df <- report$exp(
-    data.frame(lapply(df, \(x) match(x, x))) |>
+    lapply(df, \(x) match(x, x)) |>
+      data.frame() |>
       stats::setNames(column_names),
     "simplifying data types"
   )
@@ -426,6 +498,13 @@ discover <- function(
     filter_nonflat_dependencies(dependencies, detset_limit),
     column_names
   )
+}
+
+format_if_float <- function(x, digits) {
+  if ((inherits(x, "numeric") || inherits(x, "complex")))
+    format(x, digits = digits, scientific = FALSE)
+  else
+    x
 }
 
 find_LHSs_dfd <- function(