+++ tests/testthat/test-discover.r
@@ -380,18 +380,15 @@ describe("discover", {
         8.54917749999999898591
       )
     )
-    # Checking values are considered different,
-    # since R-hub tests indicate this might differ between OSes
-    expect_false(identical(
-      8.54917750000000076227,
-      8.54917749999999898591
-    ))
-    expect_false(`==`(
-      8.54917750000000076227,
-      8.54917749999999898591
-    ))
     expect_identical(
-      discover(df, 1),
+      discover(df, 1, digits = 8),
+      functional_dependency(
+        list(list(character(), "x"), list(character(), "y")),
+        c("x", "y")
+      )
+    )
+    expect_identical(
+      discover(df, 1, digits = 20),
       functional_dependency(
         list(list(character(), "x")),
         c("x", "y")