+++ NEWS.md
@@ -1,5 +1,11 @@
 # autodb (development version)
 
+Continuing efforts to prepare for submission to CRAN.
+
+## Improvements
+
+* Changed handling for numerical/complex variables in `autodb` and `discover`, to round to a number of significant digits. Due to the nature of floating-point, and the definition of a functional dependency, floating-point values can't be compared using either equality (`==`) or `all.equal` for the purposes of functional dependency discover / validation, and have the result be consistent between different machines. Because of this, floating-point variables are now rounded to a small level of precision by default before processing. If the data frame is being loaded from a file, we recommend reading any numerical/complex variables as character values (strings), if it's appropriate, to avoid loss of precision.
+
 # autodb 2.1.1
 
 Some minor changes to documentation and tests, to allow for package updates and submission to CRAN.